indentation
I'd like to figure out a way to indent the output similarly. One option
is doing,

  heading "Creating hosts"
  echo "$(git init $HOME/sandbox/host-X.git --bare --template=
	  git init $HOME/sandbox/host-Z.git --bare --template=)" | sed 's/^/ /'

The problem with doing that is it swallows any errors thrown in the
subshell. Maybe if I could pull it into a "section" function that runs
the given commands in a subshell, uses sed to indent the output, and
actually throws an error if the subshell exits with a non-zero exit
code. So far I haven't been able to implement the part of the function
that takes all the arguments and correctly passes them as commands into
the subshell. I tried
  echo "$($@)"
and
  echo "$(${@})"
but the argument handling is getting mangled.